+++ src/main/java/org/tictoc/tictoc/domain/auction/repository/history/AuctionHistoryRepository.java
@@ -1,7 +1,7 @@
-package org.tictoc.tictoc.domain.auction.repository;
+package org.tictoc.tictoc.domain.auction.repository.history;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.tictoc.tictoc.domain.auction.entity.AuctionHistory;
+import org.tictoc.tictoc.domain.auction.entity.history.AuctionHistory;
 
 public interface AuctionHistoryRepository extends JpaRepository<AuctionHistory, Long> {
 }
